+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @varlinor/cli
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @varlinor/node-tools@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes